Basic Optic Zeroing Calculator
I started on a calculator to help you zero in optics. I figured out all of the math, assigned variable names, wrote a forumla, and created an HTML file to create a user interface (idk if thats exactly what it is but I think it is). Now I just need to figure out how to get the output from some functions into other functions and it should work. Then I can improve the UI and add more functionallity incorporating DOPE and weather conditions.