I created an app mostly on my own
I made a program to help zero in optics. I had to do some googling on the html. After hours of trying to get the Javascript right I asked chat gbt to write the program and based my javascript off of its solution. It's solution was much simpler than mine and actually worked. It's solution avoided the problem of scope very simply and easily. I wrote down all of the different calculations I had been trying to do in seperate functions into one formula and gave that to chat. If I had done that in the first place I may have come to a similar solution to chat and it probably wouldve worked.