Added a Waste Array
I added an array that accounts for waste in the stringers. I would later like to have the function calculate the materials needed by the length of each individual side of the fence. Better calculating waste on stringers as well as not overestimating the number of boards and screws needed. At that point I may be able to have the function test and see whether or not the extra needed for one short section could be taken from the waste on another section so that instead of waste stacking up on multiple short sections waste from one section would be used on another not only taking it out of the waste array but also preventing more waste from being created.