Scale Generator
I got the scale generator to generate scales based on user selected options for scale type and tonic on page 4. Now i need to build this into page one and change its functionality on this page to show the scale on the fretboard. I would like to have the option to show it on the entire fretboard or one shape at a time. the first one shouldnt be too hard but I forsee the second one being challenging.