I got the new version of stick control working
All of the major functions now work. You can stop and start it without loosing your place and you can change the bpm and note type mid workout. now i need to let you jump around to or start at different parts. There are also a few minor aestetic problems to be fixed. Namely that it doesnt show measure number 1 as 1 it shows it as 40, and the first measure is fucked up.