Added functionality and made it look a bit better
Created my own photo of the staff the wrote out an array of objects containing the text for each rudiment. had JS dynamically change the rudiment text as needed. I also made the text change with the viewport to keep it sized appropriately to the image. I then made all of the images in the side bar clickable. whichever one you click will start the workout on that rudiment